be/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-decouple
django-allauth
requests
channels
